+++ package.json
@@ -8,13 +8,14 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "vitest"
   },
   "keywords": [],
   "author": "TAServers",
   "license": "MIT",
   "devDependencies": {
-    "prettier": "^3.5.3"
+    "prettier": "^3.5.3",
+    "vitest": "^3.1.3"
   },
   "dependencies": {
     "@actions/cache": "^4.0.3",
